RLI_agent/pyproject.toml
[project]
name = "rli-agent"
version = "0.1.0"
description = "Multi-agent RLI task delivery system using Claude Agent SDK"
requires-python = ">=3.10"
dependencies = [
    "claude-code-sdk>=0.0.20",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "google-genai>=1.0.0",
    "firebase-admin>=6.0.0",
    "websockets>=12.0",
    "fal-client>=0.4.0",
    "runwayml>=0.9.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["rli_agent*"]

[project.scripts]
rli-agent = "rli_agent.agent:main"
rli-server = "rli_agent.server:main"
